Task

+++ Task-2/posi-neg-0.cpp
@@ -2,17 +2,14 @@
 using namespace std;
  
 int main() {
-    double n;
- 
+     float  n;
     cout << "Enter a number: ";
     cin >> n;
- 
     if (n > 0)
         cout << n << " is positive." << endl;
     else if (n < 0)
         cout << n << " is negative." << endl;
     else
         cout << "The number is zero." << endl;
- 
     return 0;
 }
